+++ apps/3000-home/next.config.js
@@ -46,6 +46,9 @@ const nextConfig = {
             requiredVersion: '5.18.3',
             version: '5.18.3',
           },
+          '@ant-design/': {
+            singleton: true,
+          },
         },
         extraOptions: {
           debug: false,