+++ apps/3000-home/pages/_app.tsx
@@ -39,7 +39,7 @@ function MyApp(props) {
   }, []);
   return (
     <StyleProvider layer>
-      <ConfigProvider prefixCls="nxt" theme={{ hashed: false }}>
+      <ConfigProvider theme={{ hashed: false }}>
         <Layout style={{ minHeight: '100vh' }} prefixCls={'dd'}>
           <React.Suspense>
             <SharedNav />