+++ apps/3001-shop/next.config.js
@@ -38,6 +38,9 @@ const nextConfig = {
             requiredVersion: '5.18.3',
             version: '5.18.3',
           },
+          '@ant-design/': {
+            singleton: true,
+          },
         },
         extraOptions: {
           exposePages: true,