+++ apps/3001-shop/pages/_app.tsx
@@ -39,7 +39,7 @@ function MyApp({ Component, pageProps }) {
 
   return (
     <StyleProvider layer>
-      <ConfigProvider prefixCls="nxt" theme={{ hashed: false }}>
+      <ConfigProvider theme={{ hashed: false }}>
         <Layout style={{ minHeight: '100vh' }}>
           <Suspense fallback={'loading'}>
             <SharedNav />