+++ apps/esbuild/libs/shared-lib/index.ts
@@ -1,0 +1,9 @@
+var _data = '';
+
+export function setData(data: string): void {
+  _data = data;
+}
+
+export function getData(): string {
+  return _data;
+}