+++ packages/sdk/src/node.ts
@@ -11,6 +11,33 @@ function importNodeModule<T>(name: string): Promise<T> {
     });
 }
 
+const loadNodeFetch = async (): Promise<typeof fetch> => {
+  const fetchModule =
+    await importNodeModule<typeof import('node-fetch')>('node-fetch');
+  return (fetchModule.default || fetchModule) as unknown as typeof fetch;
+};
+
+const lazyLoaderHookFetch = async (
+  input: RequestInfo | URL,
+  init?: RequestInit,
+): Promise<Response> => {
+  // @ts-ignore
+  const loaderHooks = __webpack_require__.federation.instance.loaderHook;
+
+  const hook = (url: RequestInfo | URL, init: RequestInit) => {
+    return loaderHooks.lifecycle.fetch.emit(url, init);
+  };
+
+  const res = await hook(input, init || {});
+  if (!res || !(res instanceof Response)) {
+    const fetchFunction =
+      typeof fetch === 'undefined' ? await loadNodeFetch() : fetch;
+    return fetchFunction(input, init || {});
+  }
+
+  return res;
+};
+
 export function createScriptNode(
   url: string,
   cb: (error?: Error, scriptContext?: any) => void,
@@ -32,66 +59,78 @@ export function createScriptNode(
     cb(new Error(`Invalid URL: ${e}`));
     return;
   }
-  const getFetch = async () => {
-    if (typeof fetch === 'undefined') {
-      const fetchModule = await importNodeModule('node-fetch');
-      //@ts-ignore
-      return fetchModule?.default || fetchModule;
-    } else {
-      return fetch;
+
+  const getFetch = async (): Promise<typeof fetch> => {
+    //@ts-ignore
+    if (typeof __webpack_require__ !== 'undefined') {
+      try {
+        //@ts-ignore
+        const loaderHooks = __webpack_require__.federation.instance.loaderHook;
+        if (loaderHooks.lifecycle.fetch) {
+          return lazyLoaderHookFetch;
+        }
+      } catch (e) {
+        console.warn(
+          'federation.instance.loaderHook.lifecycle.fetch failed:',
+          e,
+        );
+      }
+    }
+
+    return typeof fetch === 'undefined' ? loadNodeFetch() : fetch;
+  };
+
+  const handleScriptFetch = async (f: typeof fetch, urlObj: URL) => {
+    try {
+      const res = await f(urlObj.href);
+      const data = await res.text();
+      const [path, vm] = await Promise.all([
+        importNodeModule<typeof import('path')>('path'),
+        importNodeModule<typeof import('vm')>('vm'),
+      ]);
+      const scriptContext = { exports: {}, module: { exports: {} } };
+      const urlDirname = urlObj.pathname.split('/').slice(0, -1).join('/');
+      const filename = path.basename(urlObj.pathname);
+
+      const script = new vm.Script(
+        `(function(exports, module, require, __dirname, __filename) {${data}\n})`,
+        filename,
+      );
+
+      script.runInThisContext()(
+        scriptContext.exports,
+        scriptContext.module,
+        eval('require'),
+        urlDirname,
+        filename,
+      );
+
+      const exportedInterface: Record<string, any> =
+        scriptContext.module.exports || scriptContext.exports;
+
+      if (attrs && exportedInterface && attrs['globalName']) {
+        const container =
+          exportedInterface[attrs['globalName']] || exportedInterface;
+        cb(undefined, container as keyof typeof scriptContext.module.exports);
+        return;
+      }
+
+      cb(
+        undefined,
+        exportedInterface as keyof typeof scriptContext.module.exports,
+      );
+    } catch (e) {
+      cb(new Error(`Script execution error: ${e}`));
     }
   };
-  console.log('fetching', urlObj.href);
-  getFetch().then((f) => {
-    f(urlObj.href)
-      .then((res: Response) => res.text())
-      .then(async (data: string) => {
-        const [path, vm]: [typeof import('path'), typeof import('vm')] =
-          await Promise.all([
-            importNodeModule<typeof import('path')>('path'),
-            importNodeModule<typeof import('vm')>('vm'),
-          ]);
-        const scriptContext = { exports: {}, module: { exports: {} } };
-        const urlDirname = urlObj.pathname.split('/').slice(0, -1).join('/');
-        const filename = path.basename(urlObj.pathname);
-        try {
-          const script = new vm.Script(
-            `(function(exports, module, require, __dirname, __filename) {${data}\n})`,
-            filename,
-          );
-          script.runInThisContext()(
-            scriptContext.exports,
-            scriptContext.module,
-            eval('require'),
-            urlDirname,
-            filename,
-          );
-          const exportedInterface: Record<string, any> =
-            scriptContext.module.exports || scriptContext.exports;
-          if (attrs && exportedInterface && attrs['globalName']) {
-            const container =
-              exportedInterface[attrs['globalName']] || exportedInterface;
-            cb(
-              undefined,
-              container as keyof typeof scriptContext.module.exports,
-            );
-            return;
-          }
-          cb(
-            undefined,
-            exportedInterface as keyof typeof scriptContext.module.exports,
-          );
-        } catch (e) {
-          // console.error('Error running script:', e);
-          cb(new Error(`Script execution error: ${e}`));
-        }
-      })
-      .catch((err: Error) => {
-        // console.error('Error fetching script:', err);
-        cb(err);
-      });
-  });
+
+  getFetch()
+    .then((f) => handleScriptFetch(f, urlObj))
+    .catch((err) => {
+      cb(err);
+    });
 }
+
 export function loadScriptNode(
   url: string,
   info: {