+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/modernjs
 
+## 0.1.30
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.3
+
+## 0.1.29
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.2
+
 ## 0.1.28
 
 ### Patch Changes