+++ apps/router-demo/router-host-2000/CHANGELOG.md
@@ -1,5 +1,19 @@
 # host
 
+## 1.0.4
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.3
+- @module-federation/bridge-react@0.2.3
+
+## 1.0.3
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.2
+- @module-federation/bridge-react@0.2.2
+
 ## 1.0.2
 
 ### Patch Changes