+++ packages/dts-plugin/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @module-federation/dts-plugin
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+  - @module-federation/sdk@0.2.3
+  - @module-federation/managers@0.2.3
+  - @module-federation/third-party-dts-extractor@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- 7d09ed8: Emit DTS relative to remote entry location
+  - @module-federation/sdk@0.2.2
+  - @module-federation/managers@0.2.2
+  - @module-federation/third-party-dts-extractor@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes