+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,19 @@
 # [3.1.0](https://github.com/module-federation/core/compare/utils-3.0.2...utils-3.1.0) (2023-10-26)
 
+## 3.0.30
+
+### Patch Changes
+
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+  - @module-federation/sdk@0.2.3
+
+## 3.0.29
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.2
+
 ## 3.0.28
 
 ### Patch Changes